JAMES-2544 moving CassandraMailQueueViewConfiguration class
